requirements.txt
fastapi
uvicorn
python-telegram-bot
pydantic
google-api-python-client
yt-dlp
python-dotenv
